attendance: move attendance related output to specific renderer
